theme-template/pyproject.toml
[tool.poetry]
name = "theme-template"
version = "0.1.0"
description = ""
authors = ["jamiefeiss <jamie.feiss@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
